Updated list.html
