--- src/main/java/com/company/user/DefaultUser.java
@@ -1,33 +1,0 @@
-package com.company.user;
-
-import jakarta.annotation.PostConstruct;
-import jakarta.transaction.Transactional;
-import lombok.Data;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-import java.util.Optional;
-
-@Data
-@Service
-@Transactional
-public class DefaultUser {
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @PostConstruct
-    public void defaultUserCreation() {
-        Optional<UserEntity> adminUser = userRepository.findByUsername("admin");
-        if (adminUser.isPresent()) {
-            return;
-        }
-        UserEntity user = new UserEntity();
-        user.setUsername("admin");
-        user.setPassword(passwordEncoder.encode("super_secret_password"));
-        Role role = roleRepository.findByName("ADMIN").get();
-        user.setRoles(Collections.singletonList(role));
-        userRepository.save(user);
-    }
-}
